+++ openapi-client/c/model/send_mail_raw.h
@@ -1,7 +1,7 @@
 /*
  * send_mail_raw.h
  *
- * Raw Email Object
+ * Request body for &#x60;POST /mail/rawsend&#x60;.  Accepts a complete RFC 822 message (all headers and body) as a verbatim string.  The relay injects it without modification, preserving any existing &#x60;DKIM-Signature&#x60; header intact.  The &#x60;From&#x60;, &#x60;To&#x60;, &#x60;Cc&#x60;, and &#x60;Bcc&#x60; addresses are extracted from the message headers automatically — you do not need to specify them separately.
  */
 
 #ifndef _send_mail_raw_H_
@@ -20,14 +20,14 @@ typedef struct send_mail_raw_t send_mail_raw_t;
 
 typedef struct send_mail_raw_t {
     char *raw_email; // string
-    int id; //numeric
+    long id; //numeric
 
     int _library_owned; // Is the library responsible for freeing this object?
 } send_mail_raw_t;
 
 __attribute__((deprecated)) send_mail_raw_t *send_mail_raw_create(
     char *raw_email,
-    int id
+    long id
 );
 
 void send_mail_raw_free(send_mail_raw_t *send_mail_raw);
