+++ openapi-client/groovy/src/main/groovy/org/openapitools/model/SendMailRaw.groovy
@@ -6,8 +6,8 @@ import io.swagger.annotations.ApiModelProperty;
 
 @Canonical
 class SendMailRaw {
-    /* The entire email contents */
+    /* The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination. */
     String rawEmail
-    /* Optional order id */
-    Integer id
+    /* Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`. */
+    Long id
 }
