+++ openapi-client/kotlin/src/main/kotlin/org/openapitools/client/models/SendMailRaw.kt
@@ -20,22 +20,22 @@ import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 /**
- * Raw Email Object
+ * Request body for `POST /mail/rawsend`.  Accepts a complete RFC 822 message (all headers and body) as a verbatim string.  The relay injects it without modification, preserving any existing `DKIM-Signature` header intact.  The `From`, `To`, `Cc`, and `Bcc` addresses are extracted from the message headers automatically — you do not need to specify them separately.
  *
- * @param rawEmail The entire email contents
- * @param id Optional order id
+ * @param rawEmail The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination.
+ * @param id Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`.
  */
 
 
 data class SendMailRaw (
 
-    /* The entire email contents */
+    /* The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination. */
     @Json(name = "raw_email")
     val rawEmail: kotlin.String,
 
-    /* Optional order id */
+    /* Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`. */
     @Json(name = "id")
-    val id: kotlin.Int? = null
+    val id: kotlin.Long? = null
 
 ) {
