+++ openapi-client/scalaz/src/main/scala/org/openapitools/client/api/SendMailRaw.scala
@@ -12,10 +12,10 @@ import org.joda.time.DateTime
 import SendMailRaw._
 
 case class SendMailRaw (
-  /* The entire email contents */
+  /* The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination. */
   raw_email: String,
-/* Optional order id */
-  id: Option[Integer])
+/* Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`. */
+  id: Option[Long])
 
 object SendMailRaw {
   import DateTimeCodecs._
