+++ swagger-client/csharp-dotnet2/src/main/CsharpDotNet2/mailbaby-client-csharp-dotnet2/Model/SendMailRaw.cs
@@ -8,25 +8,25 @@ using Newtonsoft.Json;
 namespace IO.Swagger.Model {
 
   /// <summary>
-  /// Raw Email Object
+  /// Request body for &#x60;POST /mail/rawsend&#x60;.  Accepts a complete RFC 822 message (all headers and body) as a verbatim string.  The relay injects it without modification, preserving any existing &#x60;DKIM-Signature&#x60; header intact.  The &#x60;From&#x60;, &#x60;To&#x60;, &#x60;Cc&#x60;, and &#x60;Bcc&#x60; addresses are extracted from the message headers automatically — you do not need to specify them separately.
   /// </summary>
   [DataContract]
   public class SendMailRaw {
     /// <summary>
-    /// The entire email contents
+    /// The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination.
     /// </summary>
-    /// <value>The entire email contents</value>
+    /// <value>The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination.</value>
     [DataMember(Name="raw_email", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "raw_email")]
     public string RawEmail { get; set; }
 
     /// <summary>
-    /// Optional order id
+    /// Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`.
     /// </summary>
-    /// <value>Optional order id</value>
+    /// <value>Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`.</value>
     [DataMember(Name="id", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "id")]
-    public int? Id { get; set; }
+    public long? Id { get; set; }
 
 
     /// <summary>
