+++ swagger-client/scala-akka-http-server/src/main/scala/io/swagger/server/model/SendMailRaw.scala
@@ -2,15 +2,15 @@ package io.swagger.server.model
 
 
 /**
- * = Root Type for RawMail =
+ * = SendMailRaw =
  *
- * Raw Email Object
+ * Request body for `POST /mail/rawsend`.  Accepts a complete RFC 822 message (all headers and body) as a verbatim string.  The relay injects it without modification, preserving any existing `DKIM-Signature` header intact.  The `From`, `To`, `Cc`, and `Bcc` addresses are extracted from the message headers automatically — you do not need to specify them separately.
  *
- * @param raw_email The entire email contents
- * @param id Optional order id
+ * @param raw_email The complete RFC 822 email message including all headers and the body. Must include at minimum `From`, `To`, `Subject`, `Date`, and `MIME-Version` headers.  DKIM signatures in the `DKIM-Signature` header are transmitted verbatim and will pass verification at the destination.
+ * @param id Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by `GET /mail`.
  */
 case class SendMailRaw (
   raw_email: String,
-  id: Option[Int]
+  id: Option[Long]
 )
