+++ swagger-client/swift5/SwaggerClient/Classes/Swaggers/Models/SendMailRaw.swift
@@ -8,16 +8,16 @@
 import Foundation
 
 
-/** Raw Email Object */
+/** Request body for &#x60;POST /mail/rawsend&#x60;.  Accepts a complete RFC 822 message (all headers and body) as a verbatim string.  The relay injects it without modification, preserving any existing &#x60;DKIM-Signature&#x60; header intact.  The &#x60;From&#x60;, &#x60;To&#x60;, &#x60;Cc&#x60;, and &#x60;Bcc&#x60; addresses are extracted from the message headers automatically — you do not need to specify them separately. */
 
 public struct SendMailRaw: Codable {
 
-    /** The entire email contents */
+    /** The complete RFC 822 email message including all headers and the body. Must include at minimum &#x60;From&#x60;, &#x60;To&#x60;, &#x60;Subject&#x60;, &#x60;Date&#x60;, and &#x60;MIME-Version&#x60; headers.  DKIM signatures in the &#x60;DKIM-Signature&#x60; header are transmitted verbatim and will pass verification at the destination. */
     public var rawEmail: String
-    /** Optional order id */
-    public var _id: Int?
+    /** Optional numeric ID of the mail order to use for SMTP authentication. If omitted the first active order on your account is used.  Valid IDs are returned by &#x60;GET /mail&#x60;. */
+    public var _id: Int64?
 
-    public init(rawEmail: String, _id: Int? = nil) {
+    public init(rawEmail: String, _id: Int64? = nil) {
         self.rawEmail = rawEmail
         self._id = _id
     }
